# [WIP] [CI] Use Ninja to speed up builds