--- a/.travis.yml
+++ b/.travis.yml
@@ -45,7 +45,17 @@ addons:
       - libgit2
       - wget
       - r
+      - ninja
     update: true
+  apt:
+    sources:
+      - sourceline: 'deb https://apt.kitware.com/ubuntu/ bionic main'
+        key_url: 'https://apt.kitware.com/keys/kitware-archive-latest.asc'
+    packages:
+      - cmake
+      - build-essential
+      - wget
+      - ninja-build
 
 before_install:
   - source tests/travis/travis_setup_env.sh