--- a/jvm-packages/create_jni.py
+++ b/jvm-packages/create_jni.py
@@ -81,12 +81,16 @@ if __name__ == "__main__":
             if sys.platform == "win32":
                 # Force x64 build on Windows.
                 maybe_generator = ' -G"Visual Studio 14 Win64"'
+                maybe_parallel_build = ""
+            elif getattr(shutil, 'which', None) is not None and shutil.which('ninja'):
+                maybe_generator = ' -GNinja'
+                maybe_parallel_build = " -- -v"
             else:
                 maybe_generator = ""
-            if sys.platform == "linux":
-                maybe_parallel_build = " -- -j $(nproc)"
-            else:
-                maybe_parallel_build = ""
+                if sys.platform == "linux":
+                    maybe_parallel_build = " -- -j $(nproc)"
+                else:
+                    maybe_parallel_build = ""
 
             args = ["-D{0}:BOOL={1}".format(k, v) for k, v in CONFIG.items()]