--- a/tests/ci_build/build_mock_cmake.sh
+++ b/tests/ci_build/build_mock_cmake.sh
@@ -4,6 +4,6 @@ set -e
 rm -rf build
 mkdir build
 cd build
-cmake -GNinja -DRABIT_MOCK=ON -DCMAKE_VERBOSE_MAKEFILE=ON ..
+cmake -GNinja -DCMAKE_BUILD_TYPE=Release -DRABIT_MOCK=ON -DCMAKE_VERBOSE_MAKEFILE=ON ..
 ninja -v
 cd ..