--- a/tests/ci_build/doxygen.sh
+++ b/tests/ci_build/doxygen.sh
@@ -13,7 +13,7 @@ branch_name=$1
 rm -rf build
 mkdir build
 cd build
-cmake .. -DBUILD_C_DOC=ON
-make -j
+cmake .. -GNinja -DBUILD_C_DOC=ON -DCMAKE_VERBOSE_MAKEFILE=ON
+ninja -v
 
 tar cvjf ${branch_name}.tar.bz2 doc_doxygen/