--- a/tests/travis/run_test.sh
+++ b/tests/travis/run_test.sh
@@ -21,8 +21,8 @@ if [ ${TASK} == "python_test" ]; then
     # Build/test
     rm -rf build
     mkdir build && cd build
-    cmake .. -DUSE_OPENMP=ON -DCMAKE_VERBOSE_MAKEFILE=ON
-    make -j$(nproc)
+    cmake .. -GNinja -DUSE_OPENMP=ON -DCMAKE_VERBOSE_MAKEFILE=ON
+    ninja -v
 
     echo "-------------------------------"
     conda activate python3
@@ -82,8 +82,8 @@ if [ ${TASK} == "cmake_test" ]; then
     rm -rf build
     mkdir build && cd build
     PLUGINS="-DPLUGIN_LZ4=ON -DPLUGIN_DENSE_PARSER=ON"
-    cmake .. -DCMAKE_VERBOSE_MAKEFILE=ON -DGOOGLE_TEST=ON -DUSE_OPENMP=ON -DUSE_DMLC_GTEST=ON ${PLUGINS}
-    make -j$(nproc)
+    cmake .. -GNinja -DCMAKE_VERBOSE_MAKEFILE=ON -DGOOGLE_TEST=ON -DUSE_OPENMP=ON -DUSE_DMLC_GTEST=ON ${PLUGINS}
+    ninja -v
     ./testxgboost
     cd ..
     rm -rf build