According to [this talk from the last year](https://slides.com/onqtam/faster_builds), Ninja is significantly faster than MSBuild.

Let the CI run with Ninja and see how much time saving is realized.