@hcho3 2020-06-28
AppVeyor
| Job name | Before | After |
|---|---|---|
| MSVC 2015 Debug | 4 min 15 sec | **2 min 37 sec** |
| MSVC 2015 Release | 9 min 31 sec | **6 min 42 sec** |
| MinGW | 8 min 37 sec | 7 min 24 sec |
| MSVC 2015 Release for R | 9 min 26 sec | **6 min 12 sec** |

Travis CI
| Job name | Before | After |
|---|---|---|
| Linux, python_sdist_test | 4 min 32 sec | 4 min 25 sec |
| MacOS, python_test | 18 min 42 sec | 22 min 11 sec |
| MacOS, python_sdist_test | 11 min 40 sec | 10 min 42 sec |
| MacOS, cmake_test | 14 min 22 sec | **12 min 44 sec** |

Jenkins
| Job name | Before | After |
|---|---|---|
| Win64 build (MSVC 2017) | 12 min 59 sec | **9 min 30 sec** |
| Linux CPU build | 4 min 58 sec | 4 min 35 sec |
| Linux CPU build (sanitizer) | 4 min 58 sec | 4 min 4 sec |
| Linux CPU build (no OpenMP) | 3 min 51 sec | 3 min 43 sec |
| Linux CPU build (mock Rabit) | 2 min 24 sec | 2 min 17 sec |
| Linux GPU build (CUDA 10.0) | 13 min 43 sec | 13 min 14 sec |
| Linux GPU build (CUDA 10.1) | 13 min 11 sec | 13 min 28 sec |
| Linux JVM build | 11 min 41 sec | 13 min 52 sec |
| Linux Doxygen | 3 min 15 sec | 2 min 38 sec |

The conclusion is that **Ninja is significant faster than MSBuild** but not than GNU Make.